UBERF-5233 reorganize ydoc utils
