+++ packages/collaboration/src/utils/__tests__/ydoc.test.ts
@@ -13,9 +13,9 @@
 // limitations under the License.
 //
 
-import { Doc as YDoc, encodeStateAsUpdate, encodeStateVector } from 'yjs'
+import { Doc as YDoc, encodeStateVector } from 'yjs'
 
-import { yDocCopy, yDocCopyWithGC, yDocFromBuffer, yDocToBuffer } from '../ydoc'
+import { yDocFromBuffer, yDocToBuffer } from '../ydoc'
 
 describe('ydoc', () => {
   it('yDocFromBuffer converts ydoc to a buffer', async () => {
@@ -35,116 +35,4 @@ describe('ydoc', () => {
     expect(target).toBeDefined()
     expect(encodeStateVector(target)).toEqual(encodeStateVector(source))
   })
-
-  describe('yDocCopy', () => {
-    it('copies document state without gc', async () => {
-      const source = new YDoc({ gc: false })
-
-      applyGarbageCollectableChanges(source)
-
-      const target = yDocCopy(source)
-
-      expect(target.gc).toBeFalsy()
-
-      // ensure target data
-      const sourceData = source.getArray('data')
-      const targetData = target.getArray('data')
-      expect(targetData.toArray()).toEqual(expect.arrayContaining(sourceData.toArray()))
-
-      // ensure target state
-      const sourceState = encodeStateVector(source)
-      const targetState = encodeStateVector(target)
-      expect(targetState).toEqual(sourceState)
-
-      // ensure target updates the same as source regardless of gc
-      const sourceUpdate = encodeStateAsUpdate(source)
-      const targetUpdate = encodeStateAsUpdate(target)
-      expect(targetUpdate).toEqual(sourceUpdate)
-    })
-
-    it('copies document state with gc', async () => {
-      const source = new YDoc({ gc: true })
-
-      applyGarbageCollectableChanges(source)
-
-      const target = yDocCopy(source)
-
-      expect(target.gc).toBeTruthy()
-
-      // ensure target data
-      const sourceData = source.getArray('data')
-      const targetData = target.getArray('data')
-      expect(targetData.toArray()).toEqual(expect.arrayContaining(sourceData.toArray()))
-
-      // ensure target state
-      const sourceState = encodeStateVector(source)
-      const targetState = encodeStateVector(target)
-      expect(targetState).toEqual(sourceState)
-
-      // ensure target updates the same as source regardless of gc
-      const sourceUpdate = encodeStateAsUpdate(source)
-      const targetUpdate = encodeStateAsUpdate(target)
-      expect(targetUpdate).toEqual(sourceUpdate)
-    })
-  })
-
-  describe('yDocCopyWithGC', () => {
-    it('copies document state without gc', async () => {
-      const source = new YDoc({ gc: false })
-
-      applyGarbageCollectableChanges(source)
-
-      const target = yDocCopyWithGC(source)
-
-      expect(target.gc).toBeFalsy()
-
-      // ensure target data
-      const sourceData = source.getArray('data')
-      const targetData = target.getArray('data')
-      expect(targetData.toArray()).toEqual(expect.arrayContaining(sourceData.toArray()))
-
-      // ensure target state
-      const sourceState = encodeStateVector(source)
-      const targetState = encodeStateVector(target)
-      expect(targetState).toEqual(sourceState)
-
-      // ensure target updates different because source is not gc-ed
-      const sourceUpdate = encodeStateAsUpdate(source)
-      const targetUpdate = encodeStateAsUpdate(target)
-      expect(targetUpdate).not.toEqual(sourceUpdate)
-    })
-
-    it('copies document state with gc', async () => {
-      const source = new YDoc({ gc: true })
-
-      applyGarbageCollectableChanges(source)
-
-      const target = yDocCopyWithGC(source)
-
-      expect(target.gc).toBeTruthy()
-
-      // ensure target data
-      const sourceData = source.getArray('data')
-      const targetData = target.getArray('data')
-      expect(targetData.toArray()).toEqual(expect.arrayContaining(sourceData.toArray()))
-
-      // ensure target state
-      const sourceState = encodeStateVector(source)
-      const targetState = encodeStateVector(target)
-      expect(targetState).toEqual(sourceState)
-
-      // ensure target updates the same because source is gc-ed
-      const sourceUpdate = encodeStateAsUpdate(source)
-      const targetUpdate = encodeStateAsUpdate(target)
-      expect(targetUpdate).toEqual(sourceUpdate)
-    })
-  })
-
-  function applyGarbageCollectableChanges (ydoc: YDoc): void {
-    const sourceData = ydoc.getArray('data')
-    sourceData.insert(0, ['a'])
-    sourceData.insert(1, [1, 2])
-    sourceData.delete(0, 1)
-    sourceData.insert(2, [3])
-  }
 })
