+++ packages/collaboration/src/utils/ydoc.ts
@@ -31,40 +31,3 @@ export function yDocToBuffer (ydoc: YDoc): Buffer {
   const update = encodeStateAsUpdate(ydoc)
   return Buffer.from(update.buffer)
 }
-
-/**
- * Copy document content as is.
- *
- * If the source document has gc parameter enabled, then garbage
- * collection will be performed. The result document will have the same
- * gc parameter value as the source document.
- *
- * @public
- * */
-export function yDocCopy (source: YDoc): YDoc {
-  const target = new YDoc({ gc: source.gc })
-
-  const update = encodeStateAsUpdate(source)
-  applyUpdate(target, update)
-
-  return target
-}
-
-/**
- * Copy document content with garbage collecting while applying update.
- *
- * Garbage collection will be performed regardless of the gc parameter
- * in the source document. The result document will have the same gc
- * parameter value as the source document.
- *
- * @public
- * */
-export function yDocCopyWithGC (source: YDoc): YDoc {
-  const target = new YDoc({ gc: source.gc })
-
-  const gc = new YDoc({ gc: true })
-  applyUpdate(gc, encodeStateAsUpdate(source))
-  applyUpdate(target, encodeStateAsUpdate(gc))
-
-  return target
-}
